requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
black==24.10.0
python-dotenv==1.0.1